move GetV6TrainingData() and related files to new directory (#1640)
