+++ src/trainingdata/trainingdata.h
@@ -1,6 +1,6 @@
 /*
   This file is part of Leela Chess Zero.
-  Copyright (C) 2018 The LCZero Authors
+  Copyright (C) 2021 The LCZero Authors
 
   Leela Chess is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
@@ -25,13 +25,10 @@
   Program grant you additional permission to convey the resulting work.
 */
 
-#include <zlib.h>
-
-#include <fstream>
-
-#include "utils/cppattributes.h"
-
 #pragma once
+
+#include "mcts/node.h"
+#include "trainingdata/writer.h"
 
 namespace lczero {
 
@@ -90,28 +87,27 @@ static_assert(sizeof(V6TrainingData) == 8356, "Wrong struct size");
 
 #pragma pack(pop)
 
-class TrainingDataWriter {
+class V6TrainingDataArray {
  public:
-  // Creates a new file to write in data directory. It will has @game_id
-  // somewhere in the filename.
-  TrainingDataWriter(int game_id);
+  V6TrainingDataArray(FillEmptyHistory white_fill_empty_history,
+                      FillEmptyHistory black_fill_empty_history,
+                      pblczero::NetworkFormat::InputFormat input_format)
+      : fill_empty_history_{white_fill_empty_history, black_fill_empty_history},
+        input_format_(input_format) {}
 
-  ~TrainingDataWriter() {
-    if (fout_) Finalize();
-  }
+  // Add a chunk.
+  void Add(const Node* node, const PositionHistory& history, Eval best_eval,
+           Eval played_eval, bool best_is_proven, Move best_move,
+           Move played_move, const NNCacheLock& nneval);
 
-  // Writes a chunk.
-  void WriteChunk(const V6TrainingData& data);
-
-  // Flushes file and closes it.
-  void Finalize();
-
-  // Gets full filename of the file written.
-  std::string GetFileName() const { return filename_; }
+  // Writes training data to a file.
+  void Write(TrainingDataWriter* writer, GameResult result,
+             bool adjudicated) const;
 
  private:
-  std::string filename_;
-  gzFile fout_;
+  std::vector<V6TrainingData> training_data_;
+  FillEmptyHistory fill_empty_history_[2];
+  pblczero::NetworkFormat::InputFormat input_format_;
 };
 
 }  // namespace lczero
